refactor: update import paths for profile-related hooks and rename files for clarity
