+++ src/app/(private)/profile/_components/nickname-edit-modal.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useState } from 'react';
 import { toast } from 'sonner';
 import { twMerge } from 'tailwind-merge';
-import { useUpdateProfile } from '~/api';
 import { extractErrorData } from '~/api/base';
 import { TimesIcon } from '~/components/icon';
 import { CTAButton } from '~/components/ui';
+import { useUpdateProfile } from '../_hooks/useUpdateProfile';
 
 interface NicknameEditModalProps {
   isOpen: boolean;
