+++ src/app/(private)/profile/_hooks/useUpdateProfile.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import type { HttpResponse, User } from '~/api';
-import { fetcher } from '../base';
-import { queryKeys } from '../queryKey';
+import { fetcher } from '~/api/base';
+import { queryKeys } from '~/api/queryKey';
 
 interface Request {
   nickname: string;
